Separate session-ingest RPC from Hono app (db-proxy pattern)
Extract RPC methods into SessionIngestRPC WorkerEntrypoint class in its
own file, reverting index.ts to export default app. Consumers reference
the named entrypoint via wrangler.jsonc service binding config. Also
adds try/catch + logging around RPC calls in session-service.ts.
